+++ NEWS.md
@@ -2,8 +2,11 @@
 
 - new: default the code to use jsonrpc version 2.0 protocol, while allowing usage of jsonrpc 1.0 too
 
-- improved: added CI testing on php 8.4 and 8.5. Default the local testing container to using PHP 8.1 on Ubuntu Jammy
+- breaking change: when creating a Request, passing in a NULL id will now automatically generate a unique id.
+  To manually create notifications, use the new `Notification` class instead.
+  The "id" of both Request and Response objects  has been made protected. To access it, use the `id()` method.
 
+- improved: added CI testing on php 8.4 and 8.5. Default the local testing container to using PHP 8.1 on Ubuntu Jammy
 
 ## JSON-RPC for PHP version 1.0-beta2 - 2024/4/15